feat: Add learning objectives to modules
Introduces a `goals` array to the `ModuleData` interface and populates it for the "Types of Triangles" and "Sum of Internal Angles" modules. This allows for more structured definition and display of learning outcomes for each educational module.

Also includes minor styling adjustments to the `LessonView` component for improved readability of lesson content.